benchmarks are now compatible with new design

+++ src/benchmark.lua
@@ -72,7 +72,7 @@ function Benchmark:getFeatures()
 end
 
 function Benchmark:exportFeatures()
-  if settings.config.skipfeature then return end
+  if (settings.config.skipfeature or settings.config.simulate) then return end
   local file = io.open(global.featureFile, "w")
   file:write("#Feature status\n#Last update: " .. get_timestamp() .. "\n\n")
   local t = {}
@@ -150,14 +150,14 @@ function Benchmark:testFeatures()
     printlog("No features to test")
   end
   for i,feature in pairs(self.featureList) do
-    printlog("Testing feature ( " .. i .. " / " .. self.feature_count .. " ): " .. feature:getName(), global.headline1)
+    printlog("Testing feature ( " .. i .. " / " .. self.feature_count .. " ): " .. feature:getName(true), global.headline1)
     feature:runTest()
     if (not settings.config.simulate) then showIndent(feature:getStatus(), 1) end
   end
   log("Step complete")
   if (settings.config.testfeature) then
     printBar()
-    self:summary()
+    self:summary(true)
     if (self.features[settings.config.testfeature] ~= nil) then
       local state = self.features[settings.config.testfeature]:isSupported()
       self:importFeatures()
@@ -174,9 +174,9 @@ function Benchmark:prepare()
   if (settings.config.testfeature) then return end
   local testcases = {}
   for id,test in pairs(self.testcases) do
-    printlog("Preparing test ( " .. id .. " / " .. #self.testcases .. " ): " .. test:getName(), global.headline1)
+    printlog("Preparing test ( " .. id .. " / " .. #self.testcases .. " ): " .. test:getName(true), global.headline1)
     test:checkFeatures(self)
-    if (not test:doSkip() or settings.config.simulate) then
+    if (not test:isDisabled() or settings.config.simulate) then
       local files = test:getLoadGenFiles()
       for i,file in pairs(files) do
         showIndent("Copying file " .. i .. "/" .. #files .. " (" .. file .. ")", 1)
@@ -197,29 +197,23 @@ function Benchmark:run()
   if (settings.config.testfeature) then return end
   if (self:checkExit()) then printlog("No test left") end
   for id,test in pairs(self.testcases) do
-    printlog("Running test ( " .. id .. " / " .. #self.testcases .. " ): " .. test:getName(), global.headline1)
+    printlog("Running test ( " .. id .. " / " .. #self.testcases .. " ): " .. test:getName(true), global.headline1)
     test:setId(id)
     if (settings.config.verbose) then test:print() end
-    if (test:getPrepareScript()) then
-      showIndent("preparing OpenFlow script (~5 sec)", 1)
-      local cmd = CommandLine.create(settings.config.localPath .. "/" .. global.benchmarkFolder .. "/" .. test:getPrepareCommand())
-      cmd:execute(settings.config.verbose)
-    end
+
     -- configure open-flow device
-    showIndent("Configuring OpenFlow device (~" .. global.timeout .. " sec)", 1, global.headline1)
-    local path = settings.config.localPath .. "/" .. global.results .. "/test_" .. test:getId()
-    test:print(path .. "_" .. test:getName() .. ".cfg")
+    showIndent("Configuring OpenFlow device (~" .. global.timeout .. " sec)", 1, global.headline2)
+    local path = settings.config.localPath .. "/" .. global.results .. "/" .. test:getName()
     local ofDev = OpenFlowDevice.create(settings.config[global.switchIP], settings.config[global.switchPort])
     ofDev:reset()
-    local flowData = ofDev:getBenchmarkFlows(test:getName(), unpack(test:getOfArgs()))
+    local flowData = ofDev:getFlowData(test)
     ofDev:createAllFiles(flowData, path)
-    if (not settings.config.simulate) then
-      ofDev:installAllFiles(path, "_ovs.output")
-      ofDev:dumpAll(path .. ".before")
-      sleep(global.timeout)
-    end
+    ofDev:installAllFiles(path, "_ovs.output")
+    ofDev:dumpAll(path .. ".before")
+    if (not settings.config.simulate) then sleep(global.timeout) end
     
     -- start loadgen
+    print("ID: " .. test:getId())
     local duration = test:getDuration() or "?"
     duration = " (measuring for " .. duration .. " sec)"
     showIndent("Starting measurement" .. duration, 1, global.headline1)
@@ -227,8 +221,9 @@ function Benchmark:run()
     local cmd = CommandLine.getRunInstance(settings:isLocal()).create()
     cmd:addCommand("cd " .. settings:get(global.loadgenWd) .. "/MoonGen")
     cmd:addCommand("./" .. test:getLoadGen() .. " " .. test:getLgArgs())
+    cmd:print()
+    lgDump:write(cmd:execute(settings.config.verbose))
     if (not settings.config.simulate) then
-      lgDump:write(cmd:execute(settings.config.verbose))
       ofDev:dumpAll(path .. ".after")
     end
     io.close(lgDump)
@@ -241,7 +236,7 @@ end
 function Benchmark:collect()
   if (settings.config.testfeature or self:checkExit()) then return end
   for id,test in pairs(self.testcases) do
-    printlog("Collecting results ( " .. id .. " / " .. #self.testcases .. " ): " .. test:getName(), global.headline1)
+    printlog("Collecting results ( " .. id .. " / " .. #self.testcases .. " ): " .. test:getName(true), global.headline1)
     local cmd = SCPCommand.create()
     cmd:switchCopyFrom(settings:isLocal(), settings:get(global.loadgenWd) .. "/" .. global.results .."/test_" .. id .. "_*.csv", settings.config.localPath .. "/" .. global.results)
     cmd:execute(settings.config.verbose)
@@ -251,23 +246,24 @@ function Benchmark:collect()
   printBar()
 end
 
-function Benchmark:summary()
+function Benchmark:summary(featureOnly)
   if (not settings.config.skipfeature and not settings.config.simulate) then
     if (self.feature_count == 0) then
       print("No features tested")
     else
       local compliance = true;
       for i,feature in pairs(self.featureList) do
-        show(string.format("Feature:   ".. ColorCode.white .. "%-22s" .. ColorCode.normal .. "%-22s %s", feature:getName(), feature:getState()..",".. feature:getRequiredOFVersion(), feature:getStatus()))
+        show(string.format("Feature:   ".. ColorCode.white .. "%-24s" .. ColorCode.normal .. "%-24s %s", feature:getName(true), feature:getState()..",".. feature:getRequiredOFVersion(), feature:getStatus()))
         compliance = compliance and (feature:getState() == global.featureState.optional or (feature:isSupported() and feature:getState() == global.featureState.required)) 
       end
       if (not settings.config.testfeature) then
-        if (compliance) then printlog("\nTestdevice is " .. settings.config[global.ofVersion] .. " compliant!", "lgreen")
+        if (compliance) then printlog("\nTestdevice has passed all required tests for " .. settings.config[global.ofVersion], "lgreen")
         else printlog("\nTestdevice is not compliant with " .. settings.config[global.ofVersion], "lred") end
       end
       printBar()
     end
   end
+  if (featureOnly) then return end
   if (settings.config.archive == true) then
     acrhiveResults()
     printBar()
